chore(dsv2lite): tighten nccl dense exchange follow-up
